refactor: watcher into seperate modules
add src/core/watcher/
moved logging and display out of watcher
created logging and display modules
